+++ include/dbsod.h
@@ -25,7 +25,7 @@ namespace py = pybind11;
 
 py::array_t<double> dbsod(
     py::array_t<double, py::array::c_style> data,
-    std::vector<float> epsSpace,
+    std::vector<double> epsSpace,
     int minPts
 );
